requirements.txt
python-dotenv==1.0.0
langgraph==0.0.15
langchain==0.1.12
langchain-community==0.0.21
firecrawl-py==1.0.0  # Replace with actual latest version
pandas==2.2.1
pydantic==2.6.3
beautifulsoup4==4.12.3
selenium==4.18.1
requests==2.31.0
webdriver-manager==4.0.1

